% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode function}
\usage{
Mode(x)
}
\arguments{
\item{x}{the input list}
}
\value{
the mode(most appearing) number/character of input list
}
\description{
This function is used to get the mode(most appearing) number/character in the input list
}
\examples{
Mode(c("A", "B", "C", "A", "A")) # should be "A"
Mode(c(1, 2, 1, 1, 1)) # should be 1

}
