% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{mode}
\alias{mode}
\title{Create a mode model from a C++ input file}
\usage{
mode(filename, quiet = FALSE, workdir = NULL)
}
\arguments{
\item{filename}{The path to a single C++ file}

\item{quiet}{Logical, indicating if compilation messages from
\code{pkgbuild} should be displayed.  Error messages will be
displayed on compilation failure regardless of the value used.}

\item{workdir}{Optional working directory to use.  If \code{NULL}
uses a temporary directory.  By using a different directory of
your choosing you can see the generated code.}
}
\value{
A generator object based on your source files
}
\description{
Create a mode model from a C++ input file.  This function will
compile the mode support around your model and return an object
that can be used to work with the model.
}
