% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Estimate the mode.}
\usage{
Mode(x)
}
\arguments{
\item{x}{Vector to find the mode of. A numeric or character vector.}
}
\value{
The mode. A numeric or character scalar.
}
\description{
This package provides a function to estimate the mode.
}
\examples{
# to a test function
Mode(x = c(1, 1, 2, 2, 3, 4, 4, 4))
}
\keyword{Descriptive}
\keyword{mode}
\keyword{statistics,}
